requirements.txt
absl-py==1.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
arabic-reshaper==3.0.0
asn1crypto==1.5.1
astunparse==1.6.3
async-timeout==5.0.1
asyncio-mqtt==0.16.2
attrs==24.2.0
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
contourpy==1.1.0
cryptography==44.0.0
cssselect2==0.7.0
cycler==0.11.0
distro==1.9.0
exceptiongroup==1.2.2
filelock==3.12.2
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
flatbuffers==23.5.26
fonttools==4.42.0
frozenlist==1.5.0
gast==0.4.0
google-ai-generativelanguage==0.1.0
google-api-core==2.23.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-generativeai==0.1.0rc1
google-pasta==0.2.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.56.2
grpcio-status==1.48.2
h11==0.14.0
h2==4.1.0
h5py==3.9.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httpx==0.27.2
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jiter==0.7.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lxml==5.3.0
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
multidict==6.1.0
mysql-connector-python==9.0.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openai==1.54.4
opencv-python==4.8.0.76
opt-einsum==3.3.0
oscrypto==1.3.0
packaging==23.1
paho-mqtt==2.1.0
pandas==2.0.3
Pillow==10.0.0
priority==2.0.0
propcache==0.2.0
proto-plus==1.25.0
protobuf==3.20.3
psutil==5.9.5
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
PyMySQL==1.1.1
pyparsing==3.0.9
pypdf==5.1.0
pypng==0.20220715.0
python-bidi==0.6.3
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
Quart==0.19.9
reportlab==4.2.5
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.10.1
seaborn==0.12.2
sentry-sdk==1.29.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
svglib==1.5.1
sympy==1.12
taskgroup==0.0.0a4
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
thop==0.1.1.post2209072238
tinycss2==1.4.0
tomli==2.1.0
torch==2.0.1
torchvision==0.15.2
tqdm==4.66.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
ultralytics==8.3.31
ultralytics-thop==2.0.11
uritools==4.0.3
urllib3==1.26.16
webencodings==0.5.1
Werkzeug==2.3.6
wrapt==1.15.0
wsproto==1.2.0
xhtml2pdf==0.2.16
yarl==1.15.2
zipp==3.16.2
